Load dreamchip from assets
b/38149414 building with dreamchip regressed startup time by 400ms. This seems to be caused by linking against the library with cc_library, but the true reason is still unknown.

In this CL, dreamchip libraries is stored as an asset instead. During runtime the asset will be copied into the data directory then loaded.

Currently the copying happens on the main thread. Whether if it's acceptable to introduce further risk to handle this is to be discussed. Disk IO could already happen before this change.

Bug: 38149414
Test: N/A
PiperOrigin-RevId: 160444223
Change-Id: I9f098e7b79df113b2992d8590b33f423942716e4