Bump version code for Dialer v11 (again)
This CL bumps the version name of Dialer from
also bumps the version code from 14***** to 15*****.

go/dialer-v10_1 did not have a version_conf_incrementer_max_version_code set, which let it to share the version code increments between v10.1 and v11. There are version codes in v10.1 which are higher than v11 which will lead to users downgrading when we submit v10.1 to the playstore (http://cl/160476738)

To make sure users are not downgraded from v11 to v10.1 when we push to the play store, I've also set 'version_conf_incrementer_max_version_code' at go/dialer-v10_1
to 1504907.

Test: N/A
PiperOrigin-RevId: 160555974
Change-Id: I4a9e64aa2fe461019e179f544fe959b8389a02cf