Use NOTIFICATION_INCOMING_CALL_QUIET when the in call UI is visible
There's a bug in O notification framework that the HUN(heads up notification) will always show when a notification with fullScreenIntent is updated.

Previously NOTIFICATION_INCOMING_CALL_QUIET is only used when the call is already active when the new incoming call is received. NOTIFICATION_INCOMING_CALL_QUIET will post the notification to the ongoing call channel instead of the incoming call, which in theory will have lower importance and the HUN will not show. In this CL, all incoming call notification will become "quiet" if the InCallUi is visible.

Bug: 62458080
Test: StatusBarNotiferTest.java
PiperOrigin-RevId: 161029526
Change-Id: I671105ecfc102beb43e52cf4d91d55a66d31cd0e