Improve GoogleCallLogAdapterTest to better emulate reality
The tests now call through the click listener to expand instead of just calling
the expanding method directly. This was done after we had a bug come up inside
the click listener that these tests did not catch.

The bug: Lightbringer.getPackageName can return null, so when using a stub there was an NPE inside the ExpandCollapseListener.

Test: only tests
PiperOrigin-RevId: 161121777
Change-Id: Ie58a4c14a32984c183b6159db2076a96cf46b37e