Run VVM activation when dialer is updated
VVM capability can change between dialer versions. For v10 the update is done through system image, so when the user get the new dialer they are guaranteed to received a onCellServiceConnected event() after boot, and activation will be triggered. This does not apply to play store update or side load.

In this CL, activation will also be triggered when the app is updated.

Bug: 35809267
Test: PackageReplacedReceiverTest
PiperOrigin-RevId: 161592369
Change-Id: I6904ef0514f9eec8151cd28be8cf76255640b108