Fixed NPE in CameraComposerFragment.
If a user took a picture then quickly closed call composer, when the image was
ready we would try to update our view state. Since our fragment is no longer
attached, this cuases a crash when we check that getContext() is not null.

fragment.getContext is never null in robolectric.
Bug: 62687110
Test: manual, cannot test b/c we cannot take photos in Espresso and
PiperOrigin-RevId: 161600278
Change-Id: If9bd98578d221fca4bc99ff17a39f917f3a8bcca