Move CallComposerActivity's DialerExecutor to onCreate
As per the documentation, DialerExecutors.createUiTaskBuilder is meant
to be called in onCreate.

As a side effect of this change, if at some point we support sending
both an image and text at the same time, the code will not work. This is
not currently planned, so this CL stays simple, rather than adding the
complexity of plumbing the old MultimediaData through
CopyAndResizeImageWorker.

Bug: 63714992
Test: CallComposerActivityIntegrationTest, manually placed EC call
PiperOrigin-RevId: 162498272
Change-Id: I710e94284b2235684a27afbfa12ea8ddef0690de