Center quickContactBadge instead of cropping.
Use the drawable dimensions to control clipping.

Before:

Adjusting the screen size from default to largest
would trigger a view reflow. However, because
the bitmap asset was already cached, and the rounding
of the tile was based on the bitmap size, the new, larger
drawable would only undergo a partial circular crop.

Now:
We scale the image uniformly and center it. Also,
use the drawable dimensions to control clipping dimensions.

Bug: 63864703
Test: manual
PiperOrigin-RevId: 163376680
Change-Id: Ic678dff2b18d3308c859818f187a42afd6563e10