Improvements to preview scaling on IMS video calls.
This is a partial rollback of previous attempts to fix preview scaling in cl/161444534 and cl/157891853.

The scaling issue was partially due to issues beneath the application, see b/63518188.

By rolling back the previous attempts, this CL improves preview scaling across all devices and OS's (assuming O builds after Aug 3).

There are still some very slight and intermittent scaling issues which we believe to be caused by the system but this CL is strictly a huge improvement. See b/63518188 for details.

Bug: 63518188
Test: manual
PiperOrigin-RevId: 165045984
Change-Id: I8e53870077e2ae86b10cd934f9f3b4508eb60ac3