Use TelecomManager#isInManagedCall starting from O.
The TelecomManager#isInCall method returns true anytime the user is in a call.
Starting in O, the APIs include support for self-managed ConnectionServices so that other apps like Duo can tell Telecom about its calls.  So, if the user is in a Duo call, isInCall would return true.
Dialer uses this to determine whether to show the "return to call in progress" when Dialer is launched.
Instead, Dialer should use TelecomManager#isInManagedCall, which only returns true if the device is in a managed call which Dialer would know about.

Bug: 36991070
Test: none
PiperOrigin-RevId: 167200903
Change-Id: I12ac7b893dcbfa2fc842ca5ab356fbbc490a098b