Send PhoneAccountHandle with ACTION_CONFIGURE_VOICEMAIL
When opening a voicemail setting from Dialer with
ACTION_CONFIGURE_VOICEMAIL, no subId is passed.
So on a multi-SIM device, it is always configuration on the default
phone regardless of SIM.

To solve this, set the PhoneAccountHandle to intent extra for
ACTION_CONFIGURE_VOICEMAIL. In the advanced settings, obtains subId
from the PhoneAccountHandle.

Test: manual - Checked that correct PhoneAccountHandle will be passed
with ACTION_CONFIGURE_VOICEMAIL intent
Bug: 65001008

Author: xuegang.x.liu <xuegang.x.liu@sony.com>
Bug: 65001008
Test: manual - Checked that correct PhoneAccountHandle will be passed
PiperOrigin-RevId: 168019873
Change-Id: I8578dd2c4ac24e4af121a90b2b175baae2832461