Dialpad VM and + are now only accessible w/ double-tap instead of long hover.
Users using touch to explore with TalkBack on the dialpad would occasionally
call voicemail if they hovered over 1 for a long period of time. This
functionality doesn't align with keyboards, where a long hover still acts as a
normal click. This CL updates the Dialpad to align better with a keyboard and
now users can only speeddial voicemail and type a 0 with double-tap activation.

Bug: 62712616
Test: manual
PiperOrigin-RevId: 168909308
Change-Id: Ib46342f173aef7c2ef64c2b42745966f83266b98