Don't truncate long voicemail transcriptions
Also repostion contact icon (for allen, is there a bug for that?)

Originally i tried wrapping the transcription in a ScrollView,
but that didn't really work since there's no easy way to set
a max-height for it.  Since most voicemails transcriptions are
not that long, an easier solution was just to not limit the number
of lines for the transcription.

screen shots:
https://drive.google.com/open?id=0B9o_KvtLkcuINkdXeVV0U0dOZVNfdExzSDNsai1Ld2F3ZlBN
https://drive.google.com/open?id=0B9o_KvtLkcuIZTZFQzRQaW1nblBPN2JkVW1hMkR3YWl1b2JV

Bug: 64211399
Test: manual
PiperOrigin-RevId: 169425482
Change-Id: I63c9165f48d902c06e288280116cd9b95d67852c