Fixed issue where all contacts weren't beign shown.
The underlying issue was that if a query was input that filtered out all of the
on device contacts, then the device rotated, the cursor would be reloaded with
all contacts filtered and getCount == 0. This would cause the cursor to never be
set in SearchCursorManager and on device contacts would no longer be shown
because the cursor wasn't being saved.

The reason this check was here, was because cursor headers were being managed
by SearchCursorManager. Since headers were  moved into the cursors we no
longer need to set cursors to null if they are empty.

Bug: 65859191
Test: existing
PiperOrigin-RevId: 169425922
Change-Id: I0baf31c3747d679f394f3746f609b9487e52472d